social_blogging_app/requirements.txt
pydantic
slowapi
redis
crewai
crewai-tools
requests
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-core
chromadb
litellm
serper-dev
pydantic
python-dotenv
httpx
requests
pyyaml
google-generativeai
